HACK: Change liip image cache paths
I don't understand how this works, but it seems to. LiipImagine is an
image editing package, and config_prod.yml was configuring its image
cache to live under /server/../../shared/web, which is a bogus path.
It's probably very specific to the coopcycle.org install setup, but it
doesn't work in a containerized environment. I removed the setting and
now no more 500s and images show up.